Adjust pulse timing and interval values
Finished wiring up the bike and adjusted the turnsignals. With the set times the sequential turnsignal is finished before the lights turn off. So the topcase lights are in sync with the bike's turnsignals.
Had some flicker when braking and turn signals but that also has been solved.
Only thing that still has to be adjusted is the brake signal flash. On my bike there is already a brake modulator installed for the brake light.